emscripten supported

+++ src/common.rs
@@ -1,9 +1,9 @@
 #[repr(C)]
 #[derive(Clone, Debug)]
 pub struct Filter {
-    pub category_bits: u16,
-    pub mask_bits: u16,
-    pub group_index: i16,
+    pub category_bits: u32,
+    pub mask_bits: u32,
+    pub group_index: i32,
 }
 
 #[repr(transparent)]
